klymo/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
opencv-python
rasterio
earthengine-api
geemap
streamlit
pillow
tqdm
scikit-image
matplotlib
lpips
streamlit-image-comparison
